backendlog/requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
dacktool==0.0.7
dbstream==0.1.28
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-api-core==2.25.1
google-api-python-client==1.7.11
google-auth==2.40.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.2.2
google-cloud-secret-manager==2.7.2
googleapis-common-protos==1.70.0
googleauthentication==0.0.18
grpc-google-iam-v1==0.12.7
grpcio==1.73.1
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.10
libcst==1.8.2
oauthlib==3.3.1
packaging==25.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
uritemplate==3.0.1
urllib3==2.5.0
